Added Exercise controlle...Get and post Exercise methods...New return code 422...Maping from DTO to Entity...Error handling in Startup to return code 500 and the err message...

+++ WT-WebAPI/Controllers/WTUsersController.cs
@@ -1,15 +1,10 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
+﻿using System.Linq;
 using System.Threading.Tasks;
 using AutoMapper;
-using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.EntityFrameworkCore;
+using WT_WebAPI.Common;
 using WT_WebAPI.Entities;
-using WT_WebAPI.Entities.DBContext;
 using WT_WebAPI.Entities.DTO;
-using WT_WebAPI.Repository;
 using WT_WebAPI.Repository.Interfaces;
 
 namespace WT_WebAPI.Controllers
@@ -28,22 +23,28 @@ namespace WT_WebAPI.Controllers
         [HttpGet("{username}")]
         public async Task<IActionResult> GetWTUser([FromRoute] string username)
         {
+            if (string.IsNullOrEmpty(username))
+            {
+                return BadRequest();
+            }
             if (!ModelState.IsValid)
             {
-                return BadRequest(ModelState);
+                return new UnprocessableEntityObjectResult(ModelState);
             }
 
             var wTUser = await _repository.GetUserByUsername(username);
 
             if (wTUser == null)
             {
-                return NotFound();
+                return NotFound("User not found");
             }
 
             var mappedUser = Mapper.Map<WTUserDTO>(wTUser);
+
             return Ok(mappedUser);
 
         }
+
 
         [HttpGet("FullInfo/{username}")]
         public async Task<IActionResult> GetWTUserFullInfo([FromRoute] string username)
@@ -65,16 +66,23 @@ namespace WT_WebAPI.Controllers
 
         }
 
+
         [HttpPut("{username}")]
         public async Task<IActionResult> PutWTUser([FromRoute] string username, [FromBody] WTUserDTO wTUser)
         {
+            if (string.IsNullOrEmpty(username))
+            {
+                return BadRequest();
+            }
+
             if (!ModelState.IsValid)
             {
-                return BadRequest(ModelState);
+                return new UnprocessableEntityObjectResult(ModelState);
             }
 
             wTUser.Username = username;
-            var result = await _repository.UpdateUser(wTUser);
+            var wtUserEntity = Mapper.Map<WTUser>(wTUser);
+            var result = await _repository.UpdateUser(wtUserEntity);
 
             if(result == false)
             {
@@ -88,9 +96,14 @@ namespace WT_WebAPI.Controllers
         [HttpDelete("{username}")]
         public async Task<IActionResult> DeleteWTUser([FromRoute] string username)
         {
+            if (string.IsNullOrEmpty(username))
+            {
+                return BadRequest();
+            }
+
             if (!ModelState.IsValid)
             {
-                return BadRequest(ModelState);
+                return new UnprocessableEntityObjectResult(ModelState);
             }
 
             var result = await _repository.DeleteUser(username);
